+++ lib/src/clojure_lsp/feature/move_coll_entry.clj
@@ -1,5 +1,7 @@
 (ns clojure-lsp.feature.move-coll-entry
   (:require
+   [clojure-lsp.queries :as q]
+   [clojure-lsp.shared :as shared]
    [clojure.string :as string]
    [rewrite-clj.node :as n]
    [rewrite-clj.zip :as z]))
@@ -22,11 +24,11 @@
        (take-while (complement z/rightmost?))
        count))
 
-(defn newline-comment? [n]
+(defn ^:private newline-comment? [n]
   (and (n/comment? n)
        (string/ends-with? (:s n) "\n")))
 
-(defn z-take-while
+(defn ^:private z-take-while
   "Returns a sequence of locations in the direction of `f` from `zloc` that
   satisfy `p?`."
   [zloc f p?]
@@ -36,7 +38,7 @@
        (take-while (complement z/end?))
        (take-while p?)))
 
-(defn z-split-with
+(defn ^:private z-split-with
   "Like clojure.core/split-with, but for a clj-rewrite zipper. Returns two
   items, a sequence of locations to the z/right* of `zloc` that satisfy `p?`,
   and the first location that doesn't."
@@ -121,7 +123,7 @@
                              ;; TODO: handle case where comma precedes comment
                              ;; `a 1, ;; one comment`
                              ;; Conceptually, the comma should stay put while
-                             ;; the entry and comment move.
+                             ;; the pair and comment move.
                              (z/skip z/right* (comp #{:whitespace :comment}
                                                     z/tag))
                              ;; relinquish that node
@@ -142,67 +144,106 @@
                           :idx  elem-idx
                           :locs (concat prefix [elem-loc] postfix)}))))))))
 
-(defn z-cursor-position [zloc]
+(defn ^:private z-cursor-position [zloc]
   (meta (z/node zloc)))
 
-(defn ^:private move-entry-zloc
-  "Move `zloc` to direction `dir` considering multiple comments cases."
+(defn ^:private elem-index-by-cursor-position
+  "Search for an element within `elems` that was at `cursor-position`.
+
+  We compare cursor position, not full loc, because comments have been separated
+  from their newlines by `seq-elems`.
+
+  This procedure is complicated by the possiblity that the cursor is on padding
+  between elements. If this was the case, we assume the intention was to move
+  the next element."
+  [cursor-position elems]
+  (reduce (fn [best-idx {:keys [idx type locs]}]
+            (let [origin? (->> locs
+                               (some (fn [elem-loc]
+                                       (= (z-cursor-position elem-loc)
+                                          cursor-position)))
+                               boolean)]
+              (case [type origin?]
+                [:elem true]     (reduced idx)
+                [:elem false]    idx
+                [:padding true]  (reduced (inc best-idx))
+                [:padding false] best-idx)))
+          -1
+          elems))
+
+(defn ^:private elem-by-index [origin-idx elems]
+  (->> elems
+       (filter (fn [{:keys [type idx]}]
+                 (and (= :elem type)
+                      (= idx origin-idx))))
+       first))
+
+(defn ^:private split-elems-at-idx [elems split-idx]
+  (split-with (fn [{:keys [type idx]}]
+                (or (= :padding type)
+                    (< idx split-idx)))
+              elems))
+
+(defn ^:private edit-collection
+  "Put revised entries back into parent."
+  [parent-zloc swapped]
+  (z/replace parent-zloc
+             (n/replace-children (z/node parent-zloc)
+                                 (->> swapped
+                                      (mapcat :locs)
+                                      (map z/node)))))
+
+(defn ^:private fix-trailing-comment
+  "After reordering, the last component may now be a comment. We need to add a
+  newline because otherwise the closing bracket will be commented out. By
+  default this aligns the bracket with rightmost element, but this can be
+  modified by providing a `movement` from the rightmost."
+  ([parent-zloc] (fix-trailing-comment parent-zloc identity))
+  ([parent-zloc movement]
+   (let [last-zloc (-> parent-zloc z/down z/rightmost*)]
+     (if (-> last-zloc z/node n/comment?)
+       (let [col (-> parent-zloc
+                     z/down
+                     z/rightmost
+                     movement
+                     z-cursor-position
+                     :col)]
+         (-> last-zloc
+             (z/insert-space-right (dec col))
+             z/insert-newline-right
+             z/up))
+       parent-zloc))))
+
+(defn ^:private move-pair-zloc
+  "Move a pair of elements around `zloc` in direction `dir` considering multiple
+  comments cases."
   [zloc dir]
-  (let [parent-zloc     (z/up zloc)
-        cursor-position (z-cursor-position zloc)
+  (let [parent-zloc (z/up zloc)
 
         elems (seq-elems parent-zloc)
 
-        ;; Search for original cursor location within elems.
-        ;; We compare cursor position, not full loc, because comments have been
-        ;; separated from their newlines by `seq-elems`.
-        ;; This algorithm is complicated by the possiblity that the cursor is on
-        ;; padding between elements. If this was the case, we assume the
-        ;; intention was to move the next element.
-        origin-idx (reduce (fn [best-idx {:keys [idx type locs]}]
-                             (let [origin? (->> locs
-                                                (some (fn [elem-loc]
-                                                        (= (z-cursor-position elem-loc)
-                                                           cursor-position)))
-                                                boolean)]
-                               (case [type origin?]
-                                 [:elem true]     (reduced idx)
-                                 [:elem false]    idx
-                                 [:padding true]  (reduced (inc best-idx))
-                                 [:padding false] best-idx)))
-                           -1
-                           elems)
+        origin-idx (elem-index-by-cursor-position (z-cursor-position zloc) elems)
         ;; Here we begin to treat elements like pairs.
-        origin-idx (if (odd? origin-idx) ;; on value
-                     (dec origin-idx)    ;; back to key
-                     origin-idx)
+        origin-idx (cond-> origin-idx
+                     ;; if on on value, go back to key
+                     (odd? origin-idx) dec)
         dest-idx   (dir (dir origin-idx))
 
-        origin-elem (->> elems
-                         (filter (fn [{:keys [type idx]}]
-                                   (and (= :elem type)
-                                        (= idx origin-idx))))
-                         first)
-        dest-elem   (->> elems
-                         (filter (fn [{:keys [type idx]}]
-                                   (and (= :elem type)
-                                        (= idx dest-idx))))
-                         first)]
+        origin-elem (elem-by-index origin-idx elems)
+        dest-elem   (elem-by-index dest-idx elems)]
     ;; In a few cases, the simple can-move-*? heuristics return the wrong
     ;; results. This happens:
-    ;; A) When on a padding line after the first pair, and `move-up` is invoked.
+    ;; A) When on a comment after the first pair, and `move-up` is invoked.
     ;; B) When on a padding line before the last pair, and `move-down` is invoked.
     ;; Movement should not be allowed in either of these cases. Unfortunately,
-    ;; there's no quick way to know we're in this situation. But now that we
-    ;; have parsed the elems, we can detect it. It manifests as the origin-idx
-    ;; or dest-idx being out of bounds, in which case either origin-elem or
+    ;; there's no quick way to know we're in this situation. But now that we've
+    ;; parsed the elems, we can detect it. It manifests as the origin-idx or
+    ;; dest-idx being out of bounds, in which case either origin-elem or
     ;; dest-elem will be missing. We bail out now to avoid erroneous swaps.
     (when (and origin-elem dest-elem)
-      (let [earlier-idx        (min origin-idx dest-idx)
-            [before rst]       (split-with (fn [{:keys [type idx]}]
-                                             (or (= :padding type)
-                                                 (< idx earlier-idx)))
-                                           elems)
+      (let [earlier-idx  (min origin-idx dest-idx)
+            [before rst] (split-elems-at-idx elems earlier-idx)
+
             [earlier-pair rst] (split-at 3 rst) ;; key, padding, value
             [interstitial rst] (split-at 1 rst) ;; padding
             [later-pair rst]   (split-at 3 rst) ;; key, padding, value
@@ -212,42 +253,132 @@
                                 interstitial
                                 earlier-pair
                                 rst)
-            ;; Put revised entries back into parent.
-            parent-zloc (z/replace parent-zloc
-                                   (n/replace-children (z/node parent-zloc)
-                                                       (->> swapped
-                                                            (mapcat :locs)
-                                                            (map z/node))))
-
-            ;; If after reordering, the last component has become a comment,
-            ;; we need to add a newline or else the closing bracket will be
-            ;; commented out.
-            last-zloc   (-> parent-zloc z/down z/rightmost*)
-            parent-zloc (if (-> last-zloc z/node n/comment?)
-                          ;; align bracket with last key
-                          (let [last-key (-> parent-zloc z/down z/rightmost z/left)
-                                col      (:col (meta (z/node last-key)))]
-                            (-> last-zloc
-                                (z/insert-space-right (dec col))
-                                z/insert-newline-right
-                                z/up))
-                          parent-zloc)]
+            parent-zloc (-> parent-zloc
+                            (edit-collection swapped)
+                            ;; align with last key, not value
+                            (fix-trailing-comment z/left))]
         [parent-zloc (first (:locs dest-elem))]))))
 
-(defn ^:private can-move-entry? [zloc]
-  (and (contains? #{:map :vector} (some-> zloc z/up z/tag))
-       (even? (count (z/child-sexprs (z/up zloc))))))
+(defn ^:private move-element-zloc
+  "Move an element around `zloc` in direction `dir` considering multiple
+  comments cases."
+  [zloc dir]
+  (let [parent-zloc (z/up zloc)
 
-(defn can-move-entry-up? [zloc]
-  (and (can-move-entry? zloc)
-       (>= (count-siblings-left zloc) 2)))
+        elems (seq-elems parent-zloc)
 
-(defn can-move-entry-down? [zloc]
-  (and (can-move-entry? zloc)
-       (>= (count-siblings-right zloc) 2)))
+        origin-idx (elem-index-by-cursor-position (z-cursor-position zloc) elems)
+        dest-idx   (dir origin-idx)
 
-(defn ^:private move-entry [zloc uri dir]
-  (when-let [[parent-loc dest-loc] (move-entry-zloc zloc dir)]
+        origin-elem (elem-by-index origin-idx elems)
+        dest-elem   (elem-by-index dest-idx elems)]
+    ;; In a few cases, the simple can-move-*? heuristics return the wrong
+    ;; results. This happens:
+    ;; A) When on a comment after the first element, and `move-up` is invoked.
+    ;; B) When on a padding line before the last element, and `move-down` is invoked.
+    ;; Movement should not be allowed in either of these cases. Unfortunately,
+    ;; there's no quick way to know we're in this situation. But now that we've
+    ;; parsed the elems, we can detect it. It manifests as the origin-idx or
+    ;; dest-idx being out of bounds, in which case either origin-elem or
+    ;; dest-elem will be missing. We bail out now to avoid erroneous swaps.
+    (when (and origin-elem dest-elem)
+      (let [earlier-idx  (min origin-idx dest-idx)
+            [before rst] (split-elems-at-idx elems earlier-idx)
+
+            [[earlier-elem padding later-elem] rst] (split-at 3 rst)
+
+            swapped (concat before
+                            [later-elem
+                             padding
+                             earlier-elem]
+                            rst)
+
+            parent-zloc (-> parent-zloc
+                            (edit-collection swapped)
+                            (fix-trailing-comment))]
+        [parent-zloc (first (:locs dest-elem))]))))
+
+(def ^:private common-binding-syms
+  "Symbols that are typically used to define bindings."
+  ;; We are not concerned with macros like `if-let` that establish one binding
+  ;; only. They can be treated like regular 'move-element' vectors.
+  #{'binding 'doseq 'for 'let 'loop 'with-local-vars 'with-open 'with-redefs})
+
+(defn ^:private establishes-bindings?
+  "Returns whether `zloc` (which should be a vector node) establishes bindings,
+  such as in `clojure.core/let`.
+
+  This uses a few heurisitics.
+
+  First, there are several clojure.core functions and macros which establish
+  bindings, such as `for`, `let`, `loop` and `binding`. It returns true if
+  `zloc` is in one of these expressions (even if the expression is actually
+  imported from some namespace besides clojure.core (TODO is it reasonable to
+  assume that any function with one of these names establishes bindings, even if
+  it isn't in clojure.core?)).
+
+  Otherwise, if the first child of `zloc` defines a local variable (according to
+  the clj-kondo diagnostics), then this also returns true. This allows library-
+  and user-defined code to declare that it establishes bindings, via `:lint-as`.
+  Originally this was the only heuristic, since it works for `let` and `for`.
+  But it fails for `bindings` and other forms where the clj-kondo analysis
+  reports that it merely references existing variables, rather than establishing
+  definitions."
+  [zloc uri {:keys [analysis]}]
+  (boolean
+    (or (when-let [outer-zloc (z/left zloc)]
+          (and (nil? (z/left outer-zloc)) ;; leftmost
+               (contains? common-binding-syms (z/sexpr outer-zloc))))
+        (let [z-pos   (z-cursor-position (z/down zloc))
+              z-scope {:name-row     (:row z-pos)
+                       :name-col     (:col z-pos)
+                       :name-end-row (:end-row z-pos)
+                       :name-end-col (:end-col z-pos)}]
+          (q/find-first (fn [element]
+                          (and (= :locals (:bucket element))
+                               (shared/inside? element z-scope)))
+                        (get analysis (shared/uri->filename uri)))))))
+
+(defn ^:private can-move-pair-up? [zloc]
+  (when (and (even? (count (z/child-sexprs (z/up zloc))))
+             (>= (count-siblings-left zloc) 2))
+    :pairwise))
+
+(defn ^:private can-move-element-up? [zloc]
+  (when (>= (count-siblings-left zloc) 1)
+    :elementwise))
+
+(defn ^:private can-move-pair-down? [zloc]
+  (when (and (even? (count (z/child-sexprs (z/up zloc))))
+             (>= (count-siblings-right zloc) 2))
+    :pairwise))
+
+(defn ^:private can-move-element-down? [zloc]
+  (when (>= (count-siblings-right zloc) 1)
+    :elementwise))
+
+(defn can-move-entry-up? [zloc uri db]
+  (let [parent-zloc (z/up zloc)]
+    (case (some-> parent-zloc z/tag)
+      :map         (can-move-pair-up? zloc)
+      :vector      (if (establishes-bindings? parent-zloc uri db)
+                     (can-move-pair-up? zloc)
+                     (can-move-element-up? zloc))
+      (:list :set) (can-move-element-up? zloc)
+      nil)))
+
+(defn can-move-entry-down? [zloc uri db]
+  (let [parent-zloc (z/up zloc)]
+    (case (some-> parent-zloc z/tag)
+      :map         (can-move-pair-down? zloc)
+      :vector      (if (establishes-bindings? parent-zloc uri db)
+                     (can-move-pair-down? zloc)
+                     (can-move-element-down? zloc))
+      (:list :set) (can-move-element-down? zloc)
+      nil)))
+
+(defn ^:private changes [uri [parent-loc dest-loc :as changed]]
+  (when changed
     {:show-document-after-edit {:uri         uri
                                 :take-focus? true
                                 :range       (z-cursor-position dest-loc)}
@@ -255,10 +386,14 @@
                                 [{:range (z-cursor-position parent-loc)
                                   :loc   parent-loc}]}}))
 
-(defn move-up [zloc uri]
-  (when (can-move-entry-up? zloc)
-    (move-entry zloc uri dec)))
+(defn move-up [zloc uri db]
+  (when-let [breadth (can-move-entry-up? zloc uri db)]
+    (changes uri (case breadth
+                   :pairwise    (move-pair-zloc zloc dec)
+                   :elementwise (move-element-zloc zloc dec)))))
 
-(defn move-down [zloc uri]
-  (when (can-move-entry-down? zloc)
-    (move-entry zloc uri inc)))
+(defn move-down [zloc uri db]
+  (when-let [breadth (can-move-entry-down? zloc uri db)]
+    (changes uri (case breadth
+                   :pairwise    (move-pair-zloc zloc inc)
+                   :elementwise (move-element-zloc zloc inc)))))